moved the initialize client to a different file
